+++ core/src/sql/hstore.rs
@@ -4,6 +4,7 @@ use std::collections::HashMap;
  * Rust type for [hstore](https://www.postgresql.org/docs/current/hstore.html).
  */
 #[derive(Clone, Debug, Default, PartialEq)]
+#[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 pub struct Hstore(HashMap<String, Option<String>>);
 
 impl Hstore {